Audio Playlist Feature Fully Added
